My backend Library code

+++ library/src/main/java/com/lms/library/Controller/AuthorController.java
@@ -1,8 +1,10 @@
 package com.lms.library.Controller;
 
-import org.springframework.beans.factory.annotation.Autowired;
+import java.util.List;
+import java.util.Optional;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.CrossOrigin;
 import org.springframework.web.bind.annotation.DeleteMapping;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
@@ -11,38 +13,84 @@ import org.springframework.web.bind.annotation.PutMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RestController;
-import com.lms.library.Model.Author;
+import com.lms.library.DTO.AuthorDTO;
 import com.lms.library.Service.AuthorService;
+import lombok.RequiredArgsConstructor;
 
+@RequiredArgsConstructor
 @RestController
 @RequestMapping("/authors")
+@CrossOrigin(origins = "http://localhost:4200")
 public class AuthorController {
-    
-    @Autowired
-    private AuthorService authorService;
+
+    private final AuthorService authorService;
 
     @GetMapping
-    public ResponseEntity<Object> getAuthors(Author authors){
-        return new ResponseEntity<>(authorService.getAuthors(authors),HttpStatus.OK);
+    public ResponseEntity<List<AuthorDTO>> getAuthors() {
+        try {
+            List<AuthorDTO> authors = authorService.getAuthors(null);
+            return new ResponseEntity<>(authors, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<Object> getAuthorById(@PathVariable int id){
-        return new ResponseEntity<>(authorService.getAuthorById(id),HttpStatus.OK);
+    public ResponseEntity<AuthorDTO> getAuthorById(@PathVariable int id) {
+        try {
+            Optional<AuthorDTO> author = authorService.getAuthorById(id);
+            if (author.isPresent()) {
+                return new ResponseEntity<>(author.get(), HttpStatus.OK);
+            } else {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PostMapping
-    public ResponseEntity<Object> addAuthor(@RequestBody Author author){
-        return new ResponseEntity<>(authorService.addAuthor(author),HttpStatus.OK);
+    public ResponseEntity<AuthorDTO> addAuthor(@RequestBody AuthorDTO authorDTO) {
+        try {
+            AuthorDTO savedAuthor = authorService.addAuthor(authorDTO);
+            return new ResponseEntity<>(savedAuthor, HttpStatus.CREATED);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @PutMapping("/{id}")
-    public ResponseEntity<Object> updateAuthor(@PathVariable int id){
-        return new ResponseEntity<>(authorService.updateAuthor(id),HttpStatus.OK);
+    public ResponseEntity<AuthorDTO> updateAuthor(@PathVariable int id, @RequestBody AuthorDTO authorDTO) {
+        try {
+            // Set the ID from the path variable to ensure we're updating the correct author
+            authorDTO.setId(id);
+            
+            // Check if author exists first
+            Optional<AuthorDTO> existingAuthor = authorService.getAuthorById(id);
+            if (!existingAuthor.isPresent()) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            
+            AuthorDTO updatedAuthor = authorService.updateAuthor(authorDTO);
+            return new ResponseEntity<>(updatedAuthor, HttpStatus.OK);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 
     @DeleteMapping("/{id}")
-    public void deleteAuthor(@PathVariable int id){
-        authorService.deleteAuthor(id);
+    public ResponseEntity<Void> deleteAuthor(@PathVariable int id) {
+        try {
+            // Check if author exists first
+            Optional<AuthorDTO> existingAuthor = authorService.getAuthorById(id);
+            if (!existingAuthor.isPresent()) {
+                return new ResponseEntity<>(HttpStatus.NOT_FOUND);
+            }
+            
+            authorService.deleteAuthor(id);
+            return new ResponseEntity<>(HttpStatus.NO_CONTENT);
+        } catch (Exception e) {
+            return new ResponseEntity<>(HttpStatus.INTERNAL_SERVER_ERROR);
+        }
     }
 }
